- Completed first working implementation of REST API using Django Piston - Simplified URL hierarchy to only using numeric surrogate keys - Changed public key URL to /keys/publickey.gpg - Fixed numerous bugs with REST API client

+++ src/server/api/handlers.py
@@ -9,49 +9,7 @@ class BaseAptRepoHandler(BaseHandler):
     """
     Abstract base class for all REST API handlers
     """
-    
-    def _find_distribution(self, distribution_id):
-        """
-        Searches for a distribution, first by the primary key and then by the name
-        """
-        distribution_model = server.aptrepo.models.Distribution
-        distributions = distribution_model.objects.filter(id=distribution_id)
-        if distributions.count() == 0:
-            distributions = distribution_model.objects.filter(name=distribution_id)
-        
-        if distributions.count() == 0:
-            raise ObjectDoesNotExist
-        
-    def _find_section(self, distribution_id, section_id):
-        """
-        Searches for a section, first by primary key and then by the name
-        """
-        section_model = server.aptrepo.models.Section
-        if distribution_id:
-            distribution = self._find_distribution(distribution_id)
-            sections = section_model.objects.filter(id=section_id, distribution=distribution)
-            if sections.count() == 0:
-                sections = section_model.objects.filter(name=section_id, distribution=distribution)
-        else:
-            sections = section_model.objects.filter(id=section_id)
-            
-        if sections.count() == 0:
-            raise ObjectDoesNotExist
-
-    def _find_package_instance(self, distribution_id, section_id, 
-                               instance_id=None, 
-                               package_name=None, version=None, architecture=None):
-        """
-        Searches for a package instance by unique ID or the (name, version, architecture) tuple
-        """
-        if instance_id:
-            return self.model.objects.get(id=instance_id)
-        else:
-            section  = self._find_section(distribution_id, section_id)
-            return self.model.objects.get(section=section,
-                                     package__name=package_name,
-                                     package__version=version,
-                                     package__architecture=architecture)
+    exclude = ()
 
 
 class PackageHandler(BaseAptRepoHandler):
@@ -62,7 +20,10 @@ class PackageHandler(BaseAptRepoHandler):
     model = server.aptrepo.models.Package
     
     def read(self, request, id=None, name=None, version=None, architecture=None):
-        return self._find_package(id, name, version, architecture)
+        try:
+            return self._find_package(id, name, version, architecture)
+        except ObjectDoesNotExist:
+            return rc.NOT_FOUND 
 
     def delete(self, request, id=None, name=None, version=None, architecture=None):
         try:
@@ -76,10 +37,10 @@ class PackageHandler(BaseAptRepoHandler):
         
     def _find_package(self, id=None, name=None, version=None, architecture=None):
         if id:
-            return self.models.objects.get(id)
+            return self.models.objects.get(id=id)
         elif name and version and architecture:
             return self.models.objects.get(package_name=name, version=version, 
-                                    architecture=architecture)
+                                           architecture=architecture)
         
 class DistributionHandler(BaseAptRepoHandler):
     """
@@ -91,8 +52,7 @@ class DistributionHandler(BaseAptRepoHandler):
     def read(self, request, distribution_id=None):
         if distribution_id:
             try:
-                return self._find_distribution(distribution_id)
-            
+                return self.model.objects.get(id=distribution_id)
             except ObjectDoesNotExist:
                 resp = rc.NOT_FOUND
                 resp.write('Distribution not found: ' + id)
@@ -109,20 +69,27 @@ class SectionHandler(BaseAptRepoHandler):
     model = server.aptrepo.models.Section
 
     def read(self, request, distribution_id=None, section_id=None):
+        # return a specific section
         if section_id:
             try:
-                return self._find_section(distribution_id, section_id)
+                return self.model.objects.get(id=section_id)
             except ObjectDoesNotExist:
                 resp = rc.NOT_FOUND
-                resp.write('Distribution not found:  ({0}:{1})'.format(
-                    distribution_id, section_id))
+                resp.write('Section not found: {0}'.format(section_id))
                 return resp
+            
+        # return all sections within a distribution
+        elif distribution_id:
+            try:
+                return self.model.objects.filter(distribution__id=distribution_id)
+            except ObjectDoesNotExist:
+                resp = rc.NOT_FOUND
+                resp.write('Distribution not found: ' + distribution_id)
+                return resp
+        
+        # return all available sections
         else:
-            if distribution_id:
-                distribution = self._find_distribution(distribution_id)
-                return self.model.objects.filter(distribution=distribution)
-            else:
-                return self.model.objects.all()
+            return self.model.objects.all()
 
 
 class PackageInstanceHandler(BaseAptRepoHandler):
@@ -132,46 +99,66 @@ class PackageInstanceHandler(BaseAptRepoHandler):
     allowed_methods=('GET', 'DELETE', 'POST')
     model = server.aptrepo.models.PackageInstance
     
-    def read(self, request, distribution_id, section_id, 
-             instance_id=None, 
+    def read(self, request, instance_id=None, section_id=None, 
              package_name=None, version=None, architecture=None):
         
-        try:
-            return self._find_package_instance(distribution_id, section_id, 
-                                               instance_id, 
-                                               package_name, version, architecture)
-        except ObjectDoesNotExist:
-            resp = rc.NOT_FOUND
-            resp.write('Package not found in section: ({0}:{1})'.format(
-                distribution_id, section_id))
-            return resp
+        # search for a specific instance
+        if instance_id or (section_id and package_name and version and architecture):
+            try:
+                return self._find_package_instance(instance_id, section_id, 
+                                                   package_name, version, architecture)
+            except ObjectDoesNotExist:
+                resp = rc.NOT_FOUND
+                if instance_id:
+                    resp.write('Package instance not found: {0}'.format(instance_id))
+                else:
+                    resp.write('Package instance not found in section {0} \
+                                matching criteria: ({1},{2},{3})'.format(section_id, 
+                                                                         package_name, 
+                                                                         version)) 
+                return resp
+            
+        # return all packages in a section
+        elif section_id:
+            return self.model.objects.filter(section__id=section_id)
+        # show that displaying all package instances in the entire repository
+        # is disallowed
+        else:
+            return rc.FORBIDDEN      
+            
                             
-    def delete(self, request, distribution_id, section_id, 
-               instance_id=None, 
+    def delete(self, request, instance_id=None, section_id=None, 
                package_name=None, version=None, architecture=None):
-        try:
-            package_instance=self._find_package_instance(distribution_id, section_id, 
-                                                         instance_id, 
-                                                         package_name, version, architecture)
-            repository = Repository()
-            repository.remove_package(package_instance.id)
-            return rc.DELETED
-        
-        except ObjectDoesNotExist:
-            return rc.NOT_FOUND
+        # delete a specific instance
+        if instance_id or (section_id and package_name and version and architecture):
+            try:
+                package_instance=self._find_package_instance(instance_id, section_id, 
+                                                             package_name, version, architecture)
+                repository = Repository()
+                repository.remove_package(package_instance.id)
+                return rc.DELETED
+            
+            except ObjectDoesNotExist:
+                return rc.NOT_FOUND
+        else:
+            return rc.FORBIDDEN      
 
-    def create(self, request, distribution_id, section_id):
+    def create(self, request, section_id):
+        
+        if not section_id:
+            return rc.BAD_REQUEST
+        
         # if a file was uploaded, let the repository handle the file
         repository = Repository()
-        section = self._find_section(distribution_id, section_id)
+        section = server.aptrepo.models.Section.objects.get(id=section_id)
         new_instance_id = None
         if request.FILES['file']:
             uploaded_file = request.FILES['file']
             new_instance_id = repository.add_package(section.distribution.name, 
-                                                   section.name, 
-                                                   uploaded_file)
+                                                     section.name, 
+                                                     uploaded_file)
         # otherwise, clone based of the source package or instance ID
-        elif request.content_type:
+        else:
             clone_args = {'dest_distribution_name' : section.distribution.name,
                           'dest_section_name' : section.name }
             if request.data['source_id']:
@@ -180,7 +167,21 @@ class PackageInstanceHandler(BaseAptRepoHandler):
                 clone_args['package_id'] = request.data['package_id']
             new_instance_id = repository.clone_package(**clone_args)
 
-        return {'new_instance_id' : new_instance_id}
+        return self.model.objects.get(id=new_instance_id)
+
+    def _find_package_instance(self, instance_id=None, 
+                               section_id=None, package_name=None, 
+                               version=None, architecture=None):
+        """
+        Searches for a package instance by unique ID or the (name, version, architecture) tuple
+        """
+        if instance_id:
+            return self.model.objects.get(id=instance_id)
+        else:
+            return self.model.objects.get(section__id=section_id,
+                                          package__name=package_name,
+                                          package__version=version,
+                                          package__architecture=architecture)
 
 
 class ActionHandler(BaseAptRepoHandler):
@@ -200,13 +201,15 @@ class ActionHandler(BaseAptRepoHandler):
         
         # retrieve all actions for a section
         if section_id:
-            section = self._find_section(distribution_id, section_id)
-            action_query['section_id'] = section.id
+            action_query['section_id'] = section_id
 
         # retrieve all actions for a distribution
         elif distribution_id:
-            distribution = self._find_distribution(distribution_id)
-            action_query['distribution_id'] = distribution.id
+            action_query['distribution_id'] = distribution_id
             
         repository=Repository()
-        return repository.get_actions(*action_query)
+        try:
+            return repository.get_actions(**action_query)
+        except ObjectDoesNotExist:
+            return rc.NOT_FOUND
+
